backend/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.3
inflection==0.5.1
Markdown==3.7
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
pytz==2024.2
PyYAML==6.0.2
ruff==0.8.4
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.28.0
whitenoise==6.8.2
yamllint==1.35.1
